requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==38.0.4
Django==4.2.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-mail-queue==3.2.5
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
geoip2==4.7.0
idna==3.4
maxminddb==2.3.0
multidict==6.0.4
Pillow==9.3.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
yarl==1.9.2
